requirements.txt
playwright>=1.40.0
streamlit>=1.30.0
pyyaml>=6.0
pydantic>=2.0.0
requests>=2.31.0
